Изм 9.5 Переброс метода получения данных во ViewModel перевод в асинхрон. Возврат значений через CallBack данные запрашиваются непосредственно во viewModel (не стал пока трогать репозиторий, надеюсь не понадобится)
