+++ app/src/main/java/ru/geekbrains/androidonkotlin/hw/mymovie/ui/contacts/ContactsListFragment.kt
@@ -1,7 +1,6 @@
 package ru.geekbrains.androidonkotlin.hw.mymovie.ui.contacts
 
 import android.os.Bundle
-import android.provider.ContactsContract
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -10,6 +9,7 @@ import androidx.fragment.app.viewModels
 import androidx.recyclerview.widget.LinearLayoutManager
 import ru.geekbrains.androidonkotlin.hw.mymovie.databinding.ContactsListFragmentBinding
 import ru.geekbrains.androidonkotlin.hw.mymovie.domain.Contact
+import ru.geekbrains.androidonkotlin.hw.mymovie.domain.interfaces.CallBack
 
 class ContactsListFragment : Fragment() {
     private var adapter: ContactsListAdapter = ContactsListAdapter()
@@ -19,28 +19,6 @@ class ContactsListFragment : Fragment() {
     private var _binding: ContactsListFragmentBinding? = null
     private val binding get() = _binding!!
 
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        val contentResolver = requireContext().contentResolver
-        val cursor = contentResolver.query(
-            ContactsContract.Contacts.CONTENT_URI,
-            null,
-            null,
-            null,
-            ContactsContract.Contacts.DISPLAY_NAME + " ASC"
-        )
-
-        val contacts = mutableListOf<Contact>()
-        val safeCursor = cursor ?: return
-        while (safeCursor.moveToNext()) {
-            val displayName =
-                safeCursor.getString(safeCursor.getColumnIndex(ContactsContract.Contacts.DISPLAY_NAME))
-            contacts.add(Contact(displayName))
-        }
-        adapter.items = contacts
-        safeCursor.close()
-    }
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
@@ -52,7 +30,11 @@ class ContactsListFragment : Fragment() {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        contactsListViewModel.fetchData()
+        contactsListViewModel.fetchData(object : CallBack<List<Contact>> {
+            override fun onResult(value: List<Contact>) {
+                adapter.items = value
+            }
+        })
         val contactsListRecyclerView = binding.contactsList
         contactsListRecyclerView.adapter = adapter
         contactsListRecyclerView.layoutManager =
